feat: add a layout prop for navigators
This adds a `layout` prop to the navigators. Similar to `getLayout`/`layout.js` in
NextJS.

This lets user add a wrapper around the navigator's content. It's different from
adding a wrapper around the component containing the navigator in the sense that
components used with the `layout` will have access to the navigator's state that'd
allow them to render UI elements using that state.

Usage:

```js
const MyLayout = ({ children }) => {
  return (
    <View>
      {children}
      <MyFooter />
    </View>
  );
}

// ...

<Stack.Navigator layout={props => <MyLayout {...props} />}>
```